quarto/fiscalite/requirements-quarto.txt
# For dynamic tables from OpenFisca parameters (e.g. TVA history in indirecte.qmd)
# Also declared in ipp/pyproject.toml; use this file only if installing outside the project.
openfisca-france>=131
pyyaml>=6.0
pandas>=2.0
jupyter>=1.0
